Separa padrão atende em dois arquivos e propaga ajustes nos raspadores derivados
