+++ data_collection/gazette/spiders/rs/rs_gravatai.py
@@ -1,9 +1,9 @@
 from datetime import date
 
-from gazette.spiders.base.atende import BaseAtendeT2Spider
+from gazette.spiders.base.atende_layoutdois import BaseAtendeL2Spider
 
 
-class RsGravataiSpider(BaseAtendeT2Spider):
+class RsGravataiSpider(BaseAtendeL2Spider):
     TERRITORY_ID = "4309209"
     name = "rs_gravatai"
     start_date = date(2015, 5, 4)  # Edição 1
